test(): dry up tests some more, add generic utils file
