+++ core/src/components/toast/test/test.utils.ts
@@ -1,11 +1,5 @@
 import { newE2EPage } from '@stencil/core/testing';
-
-export function cleanScreenshotName(screenshotName: string): string {
-  return screenshotName
-    .replace(/([-])/g, ' ')
-    .replace(/[^0-9a-zA-Z\s]/gi, '')
-    .toLowerCase();
-}
+import { generateE2EUrl, cleanScreenshotName } from '../../../utils/test/utils';
 
 export async function testToast(
   type: string,
@@ -14,9 +8,8 @@ export async function testToast(
   screenshotName: string = cleanScreenshotName(selector)
 ) {
   try {
-    let pageUrl = `/src/components/toast/test/${type}?ionic:_testing=true`;
+    const pageUrl = generateE2EUrl('toast', type, rtl)
     if (rtl) {
-      pageUrl = `${pageUrl}&rtl=true`;
       screenshotName = `${screenshotName} rtl`;
     }
