+++ core/src/utils/test/utils.ts
@@ -1,0 +1,15 @@
+export function generateE2EUrl(component: string, type: string, rtl: boolean = false): string {
+  let url = `/src/components/${component}/test/${type}?ionic:_testing=true`;
+  if (rtl) {
+    url = `${url}&rtl=true`;
+  }
+  
+  return url;
+}
+
+export function cleanScreenshotName(screenshotName: string): string {
+  return screenshotName
+    .replace(/([-])/g, ' ')
+    .replace(/[^0-9a-zA-Z\s]/gi, '')
+    .toLowerCase();
+}
